+++ src/index.ts
@@ -8,4 +8,4 @@ import { RoomSensorThermostatPlatform } from './platform';
  */
 export = (api: API) => {
   api.registerPlatform(PLATFORM_NAME, RoomSensorThermostatPlatform);
-}
+};