+++ src/platform.ts
@@ -1,12 +1,26 @@
-/* eslint-disable max-len */
-import { API, DynamicPlatformPlugin, Logger, PlatformAccessory, PlatformConfig, Service, Characteristic } from 'homebridge';
+import {
+  API,
+  DynamicPlatformPlugin,
+  Logger,
+  PlatformAccessory,
+  PlatformConfig,
+  Service,
+  Characteristic,
+} from 'homebridge';
 import { interval } from 'rxjs';
 import axios, { AxiosInstance } from 'axios';
 import * as qs from 'querystring';
 import { readFileSync, writeFileSync } from 'fs';
-
-import { PLATFORM_NAME, PLUGIN_NAME, AuthURL, LocationURL, DeviceURL, UIurl } from './settings';
-import { RoomSensorThermostat } from './roomSensorThermostat';
+import {
+  PLATFORM_NAME,
+  PLUGIN_NAME,
+  AuthURL,
+  LocationURL,
+  DeviceURL,
+  UIurl,
+} from './settings';
+import { RoomSensorThermostat } from './RoomSensors/roomSensorThermostat';
+import * as configTypes from './configTypes';
 
 /**
  * HomebridgePlatform
@@ -15,7 +29,8 @@ import { RoomSensorThermostat } from './roomSensorThermostat';
  */
 export class RoomSensorThermostatPlatform implements DynamicPlatformPlugin {
   public readonly Service: typeof Service = this.api.hap.Service;
-  public readonly Characteristic: typeof Characteristic = this.api.hap.Characteristic;
+  public readonly Characteristic: typeof Characteristic = this.api.hap
+    .Characteristic;
 
   // this is used to track restored cached accessories
   public readonly accessories: PlatformAccessory[] = [];
@@ -24,7 +39,8 @@ export class RoomSensorThermostatPlatform implements DynamicPlatformPlugin {
     responseType: 'json',
   });
 
-  rooms: any;
+  locations!: configTypes.location | any;
+  sensoraccessory!: configTypes.sensoraccessory;
 
   constructor(
     public readonly log: Logger,
@@ -33,7 +49,7 @@ export class RoomSensorThermostatPlatform implements DynamicPlatformPlugin {
   ) {
     this.log.debug('Finished initializing platform:', this.config.name);
     // only load if configured
-    if (!config) {
+    if (!this.config) {
       return;
     }
 
@@ -42,13 +58,15 @@ export class RoomSensorThermostatPlatform implements DynamicPlatformPlugin {
       this.verifyConfig();
       this.log.debug('Config OK');
     } catch (e) {
-      this.log.error(e.message);
+      this.log.error(JSON.stringify(e.message));
+      this.log.debug(JSON.stringify(e));
       return;
     }
 
     // setup axios interceptor to add headers / api key to each request
     this.axios.interceptors.request.use((request) => {
-      request.headers.Authorization = 'Bearer ' + this.config.credentials.accessToken;
+      request.headers.Authorization =
+        'Bearer ' + this.config.credentials.accessToken;
       request.params = request.params || {};
       request.params.apikey = this.config.credentials.consumerKey;
       request.headers['Content-Type'] = 'application/json';
@@ -62,19 +80,35 @@ export class RoomSensorThermostatPlatform implements DynamicPlatformPlugin {
     this.api.on('didFinishLaunching', async () => {
       log.debug('Executed didFinishLaunching callback');
       // run the method to discover / register your devices as accessories
-      try {
-        await this.discoverDevices();
-      } catch (e) {
-        this.log.error('Failed to refresh access token.', e.message);
-      }
-
       interval((1800 / 3) * 1000).subscribe(async () => {
         try {
           await this.getAccessToken();
         } catch (e) {
-          this.log.error('Failed to refresh access token.');
+          this.log.error(
+            'Failed to refresh access token.',
+            JSON.stringify(e.message),
+          );
+          this.log.debug(JSON.stringify(e));
         }
       });
+      try {
+        this.locations = await this.discoverlocations();
+      } catch (e) {
+        this.log.error(
+          'Failed to Discover Locations.',
+          JSON.stringify(e.message),
+        );
+        this.log.debug(JSON.stringify(e));
+      }
+      try {
+        this.discoverDevices();
+      } catch (e) {
+        this.log.error(
+          'Failed to Discover Devices.',
+          JSON.stringify(e.message),
+        );
+        this.log.debug(JSON.stringify(e));
+      }
     });
   }
 
@@ -96,21 +130,44 @@ export class RoomSensorThermostatPlatform implements DynamicPlatformPlugin {
     if (!this.config.options || typeof this.config.options !== 'object') {
       this.config.options = {};
     }
-    if (!this.config.options.roomsensor || typeof this.config.options.roomsensor !== 'object') {
-      this.config.options.roomsensor = {};
+    if (
+      !this.config.options.thermostat ||
+      typeof this.config.options.thermostat !== 'object'
+    ) {
+      this.config.options.thermostat = {};
     }
-    // Room Sensor Config Options
-    this.config.options.roomsensor.hide_battery;
-    this.config.options.roomsensor.hide_temperature;
-    this.config.options.roomsensor.hide_occupancy;
-    this.config.options.roomsensor.hide_motion;
-    this.config.options.roomsensor.hide_humidity;
+    if (
+      !this.config.options.roompriority ||
+      typeof this.config.options.roompriority !== 'object'
+    ) {
+      this.config.options.roompriority = {};
+    }
+
+    // Thermostat Config Options
+    this.config.options.thermostat.thermostatSetpointStatus =
+      this.config.options.thermostat.thermostatSetpointStatus ||
+      'PermanentHold';
+
+    // Room Priority Config Options
+    this.config.options.roompriority.thermostat = true;
+    this.config.options.roompriority.priorityType =
+      this.config.options.roompriority.priorityType || 'PickARoom';
+
+    /**
+     * Hidden Device Discovery Option
+     * This will disable adding any device and will just output info.
+     */
+    this.config.devicediscovery;
 
     this.config.options.ttl = this.config.options.ttl || 1800; // default 1800 seconds
-    this.config.options.debug = this.config.options.debug || false; // default false
 
-    if (!this.config.credentials.consumerSecret && this.config.options.ttl < 1800) {
-      this.log.debug('TTL must be set to 1800 or higher unless you setup your own consumerSecret.');
+    if (
+      !this.config.credentials.consumerSecret &&
+      this.config.options.ttl < 1800
+    ) {
+      this.log.debug(
+        'TTL must be set to 1800 or higher unless you setup your own consumerSecret.',
+      );
       this.config.options.ttl = 1800;
     }
 
@@ -132,34 +189,40 @@ export class RoomSensorThermostatPlatform implements DynamicPlatformPlugin {
     let result: any;
 
     if (this.config.credentials.consumerSecret) {
-      result = (await axios({
-        url: AuthURL,
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/x-www-form-urlencoded',
-        },
-        auth: {
-          username: this.config.credentials.consumerKey,
-          password: this.config.credentials.consumerSecret,
-        },
-        data: qs.stringify({
-          grant_type: 'refresh_token',
-          refresh_token: this.config.credentials.refreshToken,
-        }),
-        responseType: 'json',
-      })).data;
+      result = (
+        await axios({
+          url: AuthURL,
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/x-www-form-urlencoded',
+          },
+          auth: {
+            username: this.config.credentials.consumerKey,
+            password: this.config.credentials.consumerSecret,
+          },
+          data: qs.stringify({
+            grant_type: 'refresh_token',
+            refresh_token: this.config.credentials.refreshToken,
+          }),
+          responseType: 'json',
+        })
+      ).data;
     } else {
       this.log.warn('Please re-link your account in the Homebridge UI.');
       // if no consumerSecret is defined, attempt to use the shared consumerSecret
       try {
-        result = (await axios.post(UIurl,
-          {
+        result = (
+          await axios.post(UIurl, {
             consumerKey: this.config.credentials.consumerKey,
             refresh_token: this.config.credentials.refreshToken,
-          },
-        )).data;
+          })
+        ).data;
       } catch (e) {
-        this.log.error('Failed to exchange refresh token for an access token.', e.message);
+        this.log.error(
+          'Failed to exchange refresh token for an access token.',
+          JSON.stringify(e.message),
+        );
+        this.log.debug(JSON.stringify(e));
         throw e;
       }
     }
@@ -179,7 +242,7 @@ export class RoomSensorThermostatPlatform implements DynamicPlatformPlugin {
   /**
    * The refresh token will periodically change.
    * This method saves the updated refresh token in the config.json file
-   * @param newRefreshToken 
+   * @param newRefreshToken
    */
   async updateRefreshToken(newRefreshToken: string) {
     try {
@@ -189,7 +252,9 @@ export class RoomSensorThermostatPlatform implements DynamicPlatformPlugin {
       }
 
       // load in the current config
-      const currentConfig = JSON.parse(readFileSync(this.api.user.configPath(), 'utf8'));
+      const currentConfig = JSON.parse(
+        readFileSync(this.api.user.configPath(), 'utf8'),
+      );
 
       // check the platforms section is an array before we do array things on it
       if (!Array.isArray(currentConfig.platforms)) {
@@ -197,10 +262,14 @@ export class RoomSensorThermostatPlatform implements DynamicPlatformPlugin {
       }
 
       // find this plugins current config
-      const pluginConfig = currentConfig.platforms.find(x => x.platform === PLATFORM_NAME);
+      const pluginConfig = currentConfig.platforms.find(
+        (x: { platform: string }) => x.platform === PLATFORM_NAME,
+      );
 
       if (!pluginConfig) {
-        throw new Error(`Cannot find config for ${PLATFORM_NAME} in platforms array`);
+        throw new Error(
+          `Cannot find config for ${PLATFORM_NAME} in platforms array`,
+        );
       }
 
       // check the .credentials is an object before doing object things with it
@@ -212,141 +281,357 @@ export class RoomSensorThermostatPlatform implements DynamicPlatformPlugin {
       pluginConfig.credentials.refreshToken = newRefreshToken;
 
       // save the config, ensuring we maintain pretty json
-      writeFileSync(this.api.user.configPath(), JSON.stringify(currentConfig, null, 4));
+      writeFileSync(
+        this.api.user.configPath(),
+        JSON.stringify(currentConfig, null, 4),
+      );
 
-      this.log.warn('Homebridge config.json has been updated with new refresh token.');
-
+      this.log.warn(
+        'Homebridge config.json has been updated with new refresh token.',
+      );
     } catch (e) {
-      this.log.error(`Failed to update refresh token in config: ${e.message}`);
+      this.log.error(
+        'Failed to update refresh token in config:',
+        JSON.stringify(e.message),
+      );
+      this.log.debug(JSON.stringify(e));
     }
   }
 
   /**
-   * This is an example method showing how to register discovered accessories.
-   * Accessories must only be registered once, previously created accessories
-   * must not be registered again to prevent "duplicate UUID" errors.
+   * this method discovers the Locations
    */
-  async discoverDevices() {
+  async discoverlocations() {
     // try and get the access token. If it fails stop here.
     try {
       await this.getAccessToken();
     } catch (e) {
-      this.log.error('Could not discover devices.', e.message);
+      this.log.error(
+        'Failed to refresh access token.',
+        JSON.stringify(e.message),
+      );
+      this.log.debug(JSON.stringify(e));
       return;
     }
-
-    // get the locations
     const locations = (await this.axios.get(LocationURL)).data;
+    this.log.info(`Total Locations Found: ${locations.length}`);
+    return locations;
+  }
 
-    this.log.debug(`# of Locations Found: ${locations.length}.`);
+  /**
+   * this method discovers the rooms at each location
+   */
+  public async Sensors(
+    device: configTypes.T9Thermostat,
+    group: configTypes.T9groups,
+    locationId: configTypes.location['locationID'],
+  ) {
+    return (
+      await this.axios.get(
+        `${DeviceURL}/thermostats/${device.deviceID}/group/${group.id}/rooms`,
+        {
+          params: {
+            locationId: locationId,
+          },
+        },
+      )
+    ).data;
+  }
 
-    // get the devices at each location
-    for (const location of locations) {
-      this.log.debug(`Getting devices for ${location.name}...`);
-
-      const locationId = location.locationID;
-      this.log.debug(locationId);
-      this.log.debug(location);
-      this.log.debug(`# of Thermostats Found: ${location.devices.length}.`);
-      for (const device of location.devices) {
-        this.log.debug(device);
-        this.log.debug(device.deviceID);
-        if ((device.deviceID.startsWith('LCC')) === true) {
-          for (const group of device.groups) {
-            this.log.debug(`Found ${device.groups.length} Group(s)`);
-            this.log.debug(device.groups);
-            this.log.debug(group);
-            this.log.debug(group.id);
-            for (const rooms of group.rooms) {
-              this.log.debug(`Found Room ${rooms}`);
-              this.log.debug(group.rooms);
-              this.log.debug(rooms);
-              this.rooms = rooms;
-            }
-            {
-              const accessory = (await this.axios.get(`${DeviceURL}/thermostats/${device.deviceID}/group/${group.id}/rooms`, {
-                params: {
-                  locationId: location.locationID,
-                },
-              })).data;
-              for (const roomaccessories of group.rooms) {
-                this.log.debug(`Found ${accessory.rooms.length} accessory.rooms`);
-                this.log.debug(group.rooms);
-                this.log.debug(roomaccessories);
+  /**
+   * This method is used to discover the your location and devices.
+   * Accessories are registered by either their DeviceClass, DeviceModel, or DeviceID
+   */
+  private async discoverDevices() {
+    if (this.locations) {
+      // get the devices at each location
+      for (const location of this.locations) {
+        this.log.info(`Getting devices for ${location.name}...`);
+        this.log.info(
+          `Total Devices Found at ${location.name}: ${location.devices.length}`,
+        );
+        this.log.debug(JSON.stringify(location));
+        this.locationinfo(location);
+        for (const device of location.devices) {
+          if (device.isAlive && device.deviceClass === 'Thermostat') {
+            if (device.deviceID.startsWith('LCC')) {
+              if (device.deviceModel.startsWith('T9')) {
+                this.deviceinfo(device);
+                this.log.debug(JSON.stringify(device));
+                try {
+                  this.discoverRoomSensors();
+                } catch (e) {
+                  this.log.error(
+                    'Failed to Find Room Sensors.',
+                    JSON.stringify(e.message),
+                  );
+                  this.log.debug(JSON.stringify(e));
+                }
               }
-              for (const accessories of accessory.rooms) {
-                this.log.debug(accessory.rooms);
-                this.log.debug(accessories);
-                for (const findaccessories of accessories.accessories) {
-                  this.log.debug(`Found ${accessories.accessories.length} accessories.accessories`);
-                  this.log.debug(accessories.accessories);
-                  this.log.debug(findaccessories);
-                  this.log.debug(findaccessories.accessoryAttribute.type);
+            } else {
+              this.log.info(
+                'A Device was found that isn\'t supported, Please open Feature Request Here: https://git.io/JUWN2, If you would like to see support.',
+              );
+            }
+          }
+        }
+      }
+    } else {
+      this.log.error(
+        'Failed to Discover Locations. Re-Link Your Honeywell Home Account.',
+      );
+    }
+  }
 
-                  // generate a unique id for the accessory this should be generated from
-                  // something globally unique, but constant, for example, the device serial
-                  // number or MAC address
-                  if (findaccessories.accessoryAttribute.type === 'IndoorAirSensor'
-                    && device.isAlive && device.deviceClass === 'Thermostat') {
-                    this.log.debug(`UDID: 
-                      ${accessories.name}${findaccessories.accessoryAttribute.type}${findaccessories.accessoryAttribute.serialNumber}`);
-                    // eslint-disable-next-line max-len
-                    const uuid = this.api.hap.uuid.generate(`${accessories.name}${findaccessories.accessoryAttribute.type}${findaccessories.accessoryAttribute.serialNumber}`);
-
-                    // see if an accessory with the same uuid has already been registered and restored from
-                    // the cached devices we stored in the `configureAccessory` method above
-                    const existingAccessory = this.accessories.find(accessory => accessory.UUID === uuid);
-
-                    if (existingAccessory) {
-                      // the accessory already exists
-                      if (device.isAlive) {
-                        this.log.info('Restoring existing accessory from cache:', existingAccessory.displayName);
-
-                        // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
-                        existingAccessory.context.firmwareRevision = findaccessories.accessoryAttribute.softwareRevision;
-                        this.api.updatePlatformAccessories([existingAccessory]);
-
-                        // create the accessory handler for the restored accessory
-                        // this is imported from `platformAccessory.ts`
-                        new RoomSensorThermostat(this, existingAccessory, locationId, device, findaccessories, group, this.rooms);
-                      } else if (!device.isAlive) {
-                        // remove platform accessories when no longer present
-                        this.api.unregisterPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [existingAccessory]);
-                        this.log.info('Removing existing accessory from cache:', existingAccessory.displayName);
+  private async discoverRoomSensors() {
+    // get the devices at each location
+    for (const location of this.locations) {
+      const locationId = location.locationID;
+      for (const device of location.devices) {
+        if (device.deviceID.startsWith('LCC')) {
+          if (device.deviceModel.startsWith('T9')) {
+            if (device.groups) {
+              const groups = device.groups;
+              for (const group of groups) {
+                const roomsensors = await this.Sensors(
+                  device,
+                  group,
+                  locationId,
+                );
+                if (roomsensors.rooms) {
+                  const rooms = roomsensors.rooms;
+                  this.log.debug(JSON.stringify(roomsensors));
+                  if (
+                    this.config.options.roompriority.thermostat
+                  ) {
+                    this.log.info(`Total Rooms Found: ${rooms.length}`);
+                  }
+                  for (const accessories of rooms) {
+                    if (accessories) {
+                      this.log.debug(JSON.stringify(accessories));
+                      for (const accessory of accessories.accessories) {
+                        const sensoraccessory = accessory;
+                        if (accessory.accessoryAttribute) {
+                          if (accessory.accessoryAttribute.type) {
+                            if (
+                              accessory.accessoryAttribute.type.startsWith(
+                                'IndoorAirSensor',
+                              )
+                            ) {
+                              this.log.debug(JSON.stringify(sensoraccessory));
+                              this.log.debug(
+                                JSON.stringify(
+                                  sensoraccessory.accessoryAttribute.name,
+                                ),
+                              );
+                              this.log.debug(
+                                JSON.stringify(
+                                  sensoraccessory.accessoryAttribute
+                                    .softwareRevision,
+                                ),
+                              );
+                              this.RoomSensorThermostat(
+                                device,
+                                locationId,
+                                sensoraccessory,
+                                group,
+                              );
+                            }
+                          }
+                        }
                       }
-                    } else {
-                      // the accessory does not yet exist, so we need to create it
-                      this.log.info('Adding new accessory:', `${accessories.name} Thermostat`);
-                      this.log.debug(`Registering new device: ${accessories.name} Thermostat - ${device.deviceID}`);
-
-                      // create a new accessory
-                      const accessory = new this.api.platformAccessory(`${accessories.name} Thermostat`, uuid);
-
-                      // store a copy of the device object in the `accessory.context`
-                      // the `context` property can be used to store any data about the accessory you may need
-                      accessory.context.device = device;
-                      accessory.context.firmwareRevision = findaccessories.accessoryAttribute.softwareRevision;
-
-                      // create the accessory handler for the newly create accessory
-                      // this is imported from `platformAccessory.ts`
-                      new RoomSensorThermostat(this, accessory, locationId, device, findaccessories, group, this.rooms);
-
-                      // link the accessory to your platform
-                      this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
                     }
-
-                  } else if (findaccessories.accessoryAttribute.type === 'Thermostat') {
-                    this.log.info(`${accessories.name} ${findaccessories.accessoryAttribute.type} was found.`);
-                  } else {
-                    // eslint-disable-next-line max-len
-                    this.log.info(`Ignoring device named ${accessories.name} - ${findaccessories.accessoryAttribute.type}, Alive Status: ${device.isAlive}`);
                   }
                 }
               }
             }
           }
-        } else {
-          this.log.info('A Device was found with a Device ID that didn\'t starts with LCC.');
+        }
+      }
+    }
+  }
+
+  private RoomSensorThermostat(
+    device: configTypes.T9Thermostat,
+    locationId: configTypes.location['locationID'],
+    sensoraccessory: configTypes.sensoraccessory,
+    group: configTypes.T9groups,
+  ) {
+    const uuid = this.api.hap.uuid.generate(
+      `${sensoraccessory.accessoryAttribute.name}-${sensoraccessory.accessoryAttribute.type}-${sensoraccessory.accessoryId}-RoomSensorThermostat-${device.deviceID}`,
+    );
+
+    // see if an accessory with the same uuid has already been registered and restored from
+    // the cached devices we stored in the `configureAccessory` method above
+    const existingAccessory = this.accessories.find(
+      (accessory: { UUID: any }) => accessory.UUID === uuid,
+    );
+
+    if (existingAccessory) {
+      // the accessory already exists
+      if (device.isAlive && this.config.options.roompriority.thermostat) {
+        this.log.info(
+          'Restoring existing accessory from cache:',
+          existingAccessory.displayName,
+        );
+
+        // if you need to update the accessory.context then you should run `api.updatePlatformAccessories`. eg.:
+        // existingAccessory.context.firmwareRevision = sensoraccessory.accessoryAttribute.softwareRevision;
+        // existingAccessory.context.name = sensoraccessory.accessoryAttribute.name;
+        // existingAccessory.context.type = sensoraccessory.accessoryAttribute.type;
+        // this.api.updatePlatformAccessories([existingAccessory]);
+
+        // create the accessory handler for the restored accessory
+        // this is imported from `platformAccessory.ts`
+        new RoomSensorThermostat(
+          this,
+          existingAccessory,
+          locationId,
+          device,
+          sensoraccessory,
+          group,
+        );
+        this.log.debug(
+          `Room Sensor Thermostat UDID: ${sensoraccessory.accessoryAttribute.name}-${sensoraccessory.accessoryAttribute.type}-${sensoraccessory.accessoryId}-RoomSensorThermostat-${device.deviceID}`,
+        );
+      } else if (
+        !device.isAlive ||
+        !this.config.options.roompriority.thermostat
+      ) {
+        this.unregisterPlatformAccessories(existingAccessory);
+      }
+    } else if (device.isAlive && this.config.options.roompriority.thermostat) {
+      // the accessory does not yet exist, so we need to create it
+      this.log.info(
+        'Adding new accessory:',
+        `${sensoraccessory.accessoryAttribute.name} ${sensoraccessory.accessoryAttribute.type} Thermostat`,
+      );
+      this.log.debug(
+        `Registering new device: ${sensoraccessory.accessoryAttribute.name} ${sensoraccessory.accessoryAttribute.type} Thermostat - ${device.deviceID}`,
+      );
+
+      // create a new accessory
+      const accessory = new this.api.platformAccessory(
+        `${sensoraccessory.accessoryAttribute.name} ${sensoraccessory.accessoryAttribute.type} Thermostat`,
+        uuid,
+      );
+
+      // store a copy of the device object in the `accessory.context`
+      // the `context` property can be used to store any data about the accessory you may need
+      accessory.context.firmwareRevision =
+        sensoraccessory.accessoryAttribute.softwareRevision;
+
+      // create the accessory handler for the newly create accessory
+      // this is imported from `platformAccessory.ts`
+      new RoomSensorThermostat(
+        this,
+        accessory,
+        locationId,
+        device,
+        sensoraccessory,
+        group,
+      );
+      this.log.debug(
+        `Room Sensor Thermostat UDID: ${sensoraccessory.accessoryAttribute.name}-${sensoraccessory.accessoryAttribute.type}-${sensoraccessory.accessoryId}-RoomSensorThermostat-${device.deviceID}`,
+      );
+
+      // link the accessory to your platform
+      this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [
+        accessory,
+      ]);
+    }
+  }
+
+  public unregisterPlatformAccessories(existingAccessory: PlatformAccessory) {
+    // remove platform accessories when no longer present
+    this.api.unregisterPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [
+      existingAccessory,
+    ]);
+    this.log.info(
+      'Removing existing accessory from cache:',
+      existingAccessory.displayName,
+    );
+  }
+
+  public locationinfo(location: configTypes.location) {
+    if (this.config.devicediscovery) {
+      if (location) {
+        this.log.warn(JSON.stringify(location));
+      }
+    }
+  }
+
+  public deviceinfo(device: {
+    deviceID: string;
+    deviceType: string;
+    deviceClass: string;
+    deviceModel: string;
+    priorityType: string;
+    settings: { fan: { allowedModes: string[]; changeableValues: any } };
+    inBuiltSensorState: { roomId: number; roomName: string };
+    groups: configTypes.T9Thermostat['groups'];
+  }) {
+    if (this.config.devicediscovery) {
+      this.log.warn(JSON.stringify(device));
+      if (device.deviceID) {
+        this.log.warn(JSON.stringify(device.deviceID));
+        this.log.error(`Device ID: ${device.deviceID}`);
+      }
+      if (device.deviceType) {
+        this.log.warn(JSON.stringify(device.deviceType));
+        this.log.error(`Device Type: ${device.deviceType}`);
+      }
+      if (device.deviceClass) {
+        this.log.warn(JSON.stringify(device.deviceClass));
+        this.log.error(`Device Class: ${device.deviceClass}`);
+      }
+      if (device.deviceModel) {
+        this.log.warn(JSON.stringify(device.deviceModel));
+        this.log.error(`Device Model: ${device.deviceModel}`);
+      }
+      if (device.priorityType) {
+        this.log.warn(JSON.stringify(device.priorityType));
+        this.log.error(`Device Priority Type: ${device.priorityType}`);
+      }
+      if (device.settings) {
+        this.log.warn(JSON.stringify(device.settings));
+        if (device.settings.fan) {
+          this.log.warn(JSON.stringify(device.settings.fan));
+          this.log.error(`Device Fan Settings: ${device.settings.fan}`);
+          if (device.settings.fan.allowedModes) {
+            this.log.warn(JSON.stringify(device.settings.fan.allowedModes));
+            this.log.error(
+              `Device Fan Allowed Modes: ${device.settings.fan.allowedModes}`,
+            );
+          }
+          if (device.settings.fan.changeableValues) {
+            this.log.warn(JSON.stringify(device.settings.fan.changeableValues));
+            this.log.error(
+              `Device Fan Changeable Values: ${device.settings.fan.changeableValues}`,
+            );
+          }
+        }
+      }
+      if (device.inBuiltSensorState) {
+        this.log.warn(JSON.stringify(device.inBuiltSensorState));
+        if (device.inBuiltSensorState.roomId) {
+          this.log.warn(JSON.stringify(device.inBuiltSensorState.roomId));
+          this.log.error(
+            `Device Built In Sensor Room ID: ${device.inBuiltSensorState.roomId}`,
+          );
+        }
+        if (device.inBuiltSensorState.roomName) {
+          this.log.warn(JSON.stringify(device.inBuiltSensorState.roomName));
+          this.log.error(
+            `Device Built In Sensor Room Name: ${device.inBuiltSensorState.roomName}`,
+          );
+        }
+      }
+      if (device.groups) {
+        this.log.warn(JSON.stringify(device.groups));
+
+        for (const group of device.groups) {
+          this.log.error(`Group: ${group.id}`);
         }
       }
     }